Customer Home: Release to 100% of new users
This is the first stage of removing the customer home A/B test logic.
The a/b test is checked, rathur than assigned, and if it doesn't exist
then the user is assumed to be in the `show` variant.

Only users not eligible for the checklist menu item or in the `hide`
variant will not see the customer home page.