Fix type errors
A couple of type errors snuck in, probably due to @types/lodash being
updated.

`concat` no longer accepts `undefined` as an argument, so default to
`[]` to have the same effect.

Using `_.get` is always a bit tricky in TS. In this case it couldn't
figure out that `requiredDependencies` was going to be a `string[]`
because TS can't understand lodash style object paths.
This change should have the same effect (if there's not step with that
name the statement will still default to `[]`).